This is messy, but at least you can run some of this stuff.
