+++ examples/python/s2ag_datasets/README.md
@@ -1,2 +1,9 @@
-This example interacts with [Datasets API endpoints](https://api.semanticscholar.org/api-docs/datasets)
-to explore available datasets.
+These examples interact with [Datasets API endpoints](https://api.semanticscholar.org/api-docs/datasets)
+to explore available datasets. These are not scripts, but rather example statements for interacting
+with the various API endpoints.
+
+[full-datasets.py](./full-datasets.py) shows how to find and download full datasets.
+
+[sample-datasets.py](./sample-datasets.py) shows how to find and inspect a smaller 
+sample of the full datasets to explore and get an idea of the data shape within.
+
