--- examples/python/s2ag_datasets/test
@@ -1,2 +1,0 @@
-#!/usr/bin/env bash
-python s2ag-datasets
