+++ examples/python/s2ag_datasets/test.py
@@ -1,0 +1,3 @@
+import subprocess
+
+subprocess.check_call("python sample-datasets.py", shell=True)
